Wrap/Unwrap TCP requests in RESP to allow stacking
The native protocol of Redis is RESP.

In order to be able to stack redis caches, the TCP
cache needs to be able to read and write RESP so as
to be able to function as a Redis proxy.

In this way an HTTP cache may be placed in front of
a TCP cache, and everything will work correctly.